+++ pkg/microservice/systemconfig/core/jira/service/jira.go
@@ -3,22 +3,22 @@ package service
 import (
 	"go.uber.org/zap"
 
-	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/email/repository/mongo"
 	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/jira/repository/models"
+	mongo2 "github.com/koderover/zadig/pkg/microservice/systemconfig/core/jira/repository/mongo"
 )
 
 func GeJira(_ *zap.SugaredLogger) (*models.Jira, error) {
-	return mongo.NewJiraColl().GetJira()
+	return mongo2.NewJiraColl().GetJira()
 }
 
 func CreateJira(jira *models.Jira, _ *zap.SugaredLogger) (*models.Jira, error) {
-	return mongo.NewJiraColl().AddJira(jira)
+	return mongo2.NewJiraColl().AddJira(jira)
 }
 
 func UpdateJira(jira *models.Jira, _ *zap.SugaredLogger) (*models.Jira, error) {
-	return mongo.NewJiraColl().UpdateJira(jira)
+	return mongo2.NewJiraColl().UpdateJira(jira)
 }
 
 func DeleteJira(_ *zap.SugaredLogger) error {
-	return mongo.NewJiraColl().DeleteJira()
+	return mongo2.NewJiraColl().DeleteJira()
 }
